+++ src/plugins/gui/providers/banxaProvider.ts
@@ -105,7 +105,6 @@ const asBanxaTxLimit = asObject({
 })
 
 const asBanxaPaymentType = asValue(
-  'CHECKOUTCREDIT',
   'CLEARJCNSELLFP',
   'CLEARJCNSELLSEPA',
   'CLEARJUNCTION',
@@ -118,8 +117,6 @@ const asBanxaPaymentType = asValue(
   'IDEAL',
   'MANUALPAYMENT',
   'MONOOVAPAYID',
-  'WORLDPAYAPPLE',
-  'WORLDPAYCREDIT',
   'WORLDPAYGOOGLE'
 )
 
@@ -828,7 +825,6 @@ const addToAllowedCurrencies = (
 }
 
 const typeMap: { [Payment in BanxaPaymentType]: FiatPaymentType } = {
-  CHECKOUTCREDIT: 'credit',
   CLEARJCNSELLFP: 'fasterpayments',
   CLEARJCNSELLSEPA: 'sepa',
   CLEARJUNCTION: 'sepa',
@@ -841,8 +837,6 @@ const typeMap: { [Payment in BanxaPaymentType]: FiatPaymentType } = {
   IDEAL: 'ideal',
   MANUALPAYMENT: 'turkishbank',
   MONOOVAPAYID: 'payid',
-  WORLDPAYAPPLE: 'applepay',
-  WORLDPAYCREDIT: 'credit',
   WORLDPAYGOOGLE: 'googlepay'
 }